workspace/LangGraph/requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
python-dotenv==1.0.0